requirements.txt
cloudinary==1.36.0
Django==4.2.7
django-cloudinary-storage==0.3.0
django-environ==0.11.2
Pillow==10.1.0
gunicorn
